Mac: Introduced dirty rects to IGraphicsContextWrapper. WaveFormControl updated.
- Added dirty rectangles to IGraphicsContextWrapper and all its
dependencies. This will break the build on other platforms and will be
fixed in the next commits.
- WaveFormControl now only draws the tiles in the dirty rectangle.
However this puts in front the bug in the tile selection algorithmn and
the fact that the bitmaps are not cutted correctly to fit in the dirty
zone. This will be fixed later.
- Fixed bug in PlayerPresenter where the song position would be updated
when the player is paused (causing redraw on controls when the position
doesn't have to be updated). This should drop a bit of CPU usage when
the player is paused.