iOS: Fixed several bugs in PlayerView and player initialization.
- PlayerView: Fixed bug where the player position would not be updated
when the playlist was changed; fixed bug where the position/length would
not be updated correctly when the album art is missing; fixed other
bugs.
- Player: Fixed issue loading correct buffer/update period values at
initialization. This was causing the player to skip when using the power
button. This doesn't fully fix player skipping though. Also updated
debug information for StreamProc callback.
- Added OnPlayerViewAppeared action in IPlayerView for mobile devices
(to resolve WaitingToStart player status).
- Updated other platforms after change in IPlayerView but not tested.
- Fixed bug in ShufflePlaylist where base.Clear() would not be called;
the playlist wasn't cleared correctly.