OSX: Fixed major graphics performance issue with CoreGraphics, might have fixed crash.
- Removed all SaveState/RestoreState calls in CoreGraphicsHelper and
moved them into every control. This is the proper use of state
save/restore per Apple's docs. Also fixed an issue in DrawRect; it was
calling Clip which completely broke every following calls to
CoreGraphics when not using SaveStare/RestoreState before each call in
the helper class. This makes the graphics performance easily 2-3x times
better than before if not more in certain cases. The CPU usage is way
down, and this *seems* to have fixed a random CoreGraphics crash bug
that was quite frequent when zooming wave forms. Hopefully this fix will
work as intended and will make the app performance a lot closer to WPF.
- Fixed bugs in TileHelper. GetEndDirtyTile is still buggy.
- Fixed bugs in WaveFormEngineService, simplified the rendering for now
to make it easier to fix other bugs and keep performance high.